+++ zcash_client_sqlite/src/chain.rs
@@ -501,6 +501,36 @@ mod tests {
     }
 
     #[test]
+    fn scan_cached_blocks_recovers_internal_change_for_unmined_non_expiring_sapling_spend() {
+        testing::pool::scan_cached_blocks_recovers_internal_change_for_unmined_non_expiring_spend::<
+            SaplingPoolTester,
+        >()
+    }
+
+    #[test]
+    #[cfg(feature = "orchard")]
+    fn scan_cached_blocks_recovers_internal_change_for_unmined_non_expiring_orchard_spend() {
+        testing::pool::scan_cached_blocks_recovers_internal_change_for_unmined_non_expiring_spend::<
+            OrchardPoolTester,
+        >()
+    }
+
+    #[test]
+    fn scan_cached_blocks_recovers_internal_change_after_mined_sapling_status_update() {
+        testing::pool::scan_cached_blocks_recovers_internal_change_after_mined_status_update::<
+            SaplingPoolTester,
+        >()
+    }
+
+    #[test]
+    #[cfg(feature = "orchard")]
+    fn scan_cached_blocks_recovers_internal_change_after_mined_orchard_status_update() {
+        testing::pool::scan_cached_blocks_recovers_internal_change_after_mined_status_update::<
+            OrchardPoolTester,
+        >()
+    }
+
+    #[test]
     fn scan_cached_blocks_detects_spends_out_of_order_sapling() {
         testing::pool::scan_cached_blocks_detects_spends_out_of_order::<SaplingPoolTester>()
     }